pyproject.toml
[project]
name = "product-image-edit"
version = "0.1.0"
description = "Gemini product-shot pipeline and Gemini corner logo removal"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]
